Link: Not render react router Link if there's no path. Instead, just render the contents of the Link directly.
Root: Import global styles. This was done previously by the Link, to use the RRLink style. This was wrong because the RRLink style did not exist anymore, and by removing the global import an issue was exposed where the global styles were not being imported anymore and everything looked bad.